requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
langchain==0.0.340
langchain-community==0.0.21
transformers==4.35.2
torch==2.1.1
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Development dependencies
pytest==7.4.3
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Additional utilities
python-multipart==0.0.6
aiofiles==23.2.1
typing-extensions==4.8.0
groq==0.4.1
